x86_64: misc fixes on platform enabling
* Provide x86_64 linker;
* Use Elf* structures instead of Elf32*, it covers both 64/32 bit;
* Fix several 64-bit types issues;
* Fix format of TRACE/LOG/DEBUG for 64-bit types;

+++ libc/bionic/pthread_key.cpp
@@ -239,7 +239,7 @@ void* pthread_getspecific(pthread_key_t key) {
   // to check that the key is properly allocated. If the key was not
   // allocated, the value read from the TLS should always be NULL
   // due to pthread_key_delete() clearing the values for all threads.
-  return (void *)(((unsigned *)__get_tls())[key]);
+  return (void *)(((uintptr_t *)__get_tls())[key]);
 }
 
 int pthread_setspecific(pthread_key_t key, const void* ptr) {
@@ -249,6 +249,6 @@ int pthread_setspecific(pthread_key_t key, const void* ptr) {
     return EINVAL;
   }
 
-  ((uint32_t *)__get_tls())[key] = (uint32_t)ptr;
+  ((uintptr_t *)__get_tls())[key] = (uintptr_t)ptr;
   return 0;
 }
